Add a WithValueTranslator option to Reconciller.
This lets the user inject code which produces helm values based on the fetched object itself (unlike Mapper which can only see Values).
This way the custom code can convert the object to a typed one and work with proper structs rather than a tree of maps to interface{}.

One wrinkle I'm not proud of is the conversion between chartutil.Values and
internal/values.Values but I don't think it's ugly enough to put it on the
critical path.