Replace WithValueTranslator implementation with upstream one (#23)
* Revert Translator commits.

Revert "Pass context to Translate(). (#8)"
This reverts commit c3df552af1b1114407c22f96aeb85c6311ed24ce.

Revert "Add a WithValueTranslator option to Reconciller. (#6)"
This reverts commit 88508a237f9f3752c928cd30ac8f568613ddb428.

* Add a WithValueTranslator option to Reconciller (cherry-picked from upstream PR).

A Translator is a way to produces helm values based on the fetched custom
resource itself (unlike `Mapper` which can only see `Values`).

This way the code which converts the custom resource to Helm values can first
convert an `Unstructured` into a regular struct, and then rely on Go type
safety rather than work with a tree of maps from `string` to `interface{}`.

Thanks to having access to a `Context`, the code can also safely access the
network, for example in order to retrieve other resources from the k8s cluster,
when they are referenced by the custom resource.