Do not ever update the status in case the generation has changed (#67)
Noticed this, because the status of RHACS operator includes an
"observedGeneration".
The helm-operator prior this change would happily refresh the object and
a status updater function registered by the RHACS operator would then
happily copy over the `metadata.generation` to the
`status.observedGeneration` field. This can happen, if, for example, the
user fiddles around with the CR spec while the helm-operator does its
reconciliation. In the end we would end up with an inconsistency:
updated `observedGeneration` even though that is not the generation
corresponding to the `spec` which was reconciled.

---------

Co-authored-by: Moritz Clasmeier <mclasmeier@redhat.com>